wip account details layout and forms update
